Attempt to fix images (and fonts?) on Cedar.
The image-url Sass mixin wasn't found, apparently because the assets
group isn't loaded in the production environment. So I moved sass-rails
(and the other assets group gems) into the global group.

This works in development; we'll see if it works on Heroku. I haven't
checked whether it works for fonts, but I'm optimistic.