Fixed broken date picker on IE 9 with asset pipeline.
IE 9 was having trouble with a script that should only have been loaded
in IE 6. I had accidentally tossed it into the asset pipeline's JS
bundle. Now it's just served from /public.

Bonus fix: The script tag for this was broken, so PNGs in IE 6 only
worked on the home page and other top-level URL paths.

https://www.pivotaltracker.com/story/show/30239231