Backbone: Fixed navigation on IE.
Backbone takes over navigation from the browser; in most browsers, it
uses the HTML History API, but in IE, it uses a hack involving #!
anchors in the URL.

We don't really want to take over navigation except within the
assessment, so now we limit intervention to that area.

Also, it seems like pushState: true forces the use of pushState even on
browsers that don't support it, so I guess let's not.

https://www.pivotaltracker.com/story/show/30434701