Dynamically keep "skip section" checkbox in sync.
If you reconsider a question in a skipped section, the box should be
checked. If you skip all questions in an unskipped section, the box
should be unchecked.

Also, if you check or uncheck the box, we automatically submit the form
without waiting for a button press. (We don't, however, disable all the
questions while the new page loads. We'll do it if we have to.)

https://www.pivotaltracker.com/story/show/27396271