Started writing test for automatic renewal.
We don't actually handle automatic renewal yet, so we don't have
the line that checks that Next Billing Date is updated. But we do
confirm that the start date remains unchanged. We thought there
was a bug that was updating start dates, but it was something else.
Still, good to have a test to make sure we keep doing things right.

https://www.pivotaltracker.com/story/show/60362290