Upgraded to Poltergeist 1.5.0.
Poltergeist is much faster, but we ran into some intermittent test
failures. The problem seems to be a race condition where we look
for the "Logout" link before the page has refreshed. Tried a few
things to make it work better, but it still fails sometimes. We're
going to live with it for now, and maybe try Capybara 2.2.