Upgraded cucumber and fog.
Cucumber because we wanted a bug fix (Around hooks run after
Background steps), fog because our extremely old version depended
on a specific version of multi_json that was too old for modern
Cucumber.

We'll test on staging to make sure the major upgrade to fog doesn't
break document uploads and downloads.