Fixed off-by-one-sometimes bug in "days to rebill."
It was computed by comparing a semi-imaginary "start date/time" with the
current moment in time, and then rounding down to get a number of days.
That meant if you signed up and then looked at your billing page, your
next billing date would be in about 23 hours, which we would call 0
days. Now we compare the current *date* with your start *date*, and call
it 1 day, i.e., you'll be charged tomorrow.