Fixed setting start_date to a date instead of a time.
We basically want it to be a date, but for unknown historical reasons,
the database column is a datetime, so let's keep it that way for now.

Also removed some specs for "days until rebill for a subscription with
no start date." We never display this information for a subscription
with no start date, and 0 isn't an intuitive answer (it's more like
NaN), so let's leave it unspecified for now.

https://www.pivotaltracker.com/story/show/67956928