Fixed test bug with todo reminders.
It was also kind of an application bug, but with no real consequences.

The test traveled to midnight next Tuesday (EST) but the app perceived
it as next Monday (CST) and didn't send reminders.

It doesn't really matter to the application because the reminders go
out at noon and it's not likely to get the day wrong because of an
hour or three. But now we use Time.zone.today in both test and app,
as we should.