Hard-code admin notification recipients.
Omar just wants notifications sent to admin@artsready.org. We considered
keeping the dynamic approach and adding a flag to individual admin
accounts to say whether they should receive notifications. But that
dynamism is not desired by South Arts; they are probably not going to
want to change recipients in the foreseeable future.

https://www.pivotaltracker.com/story/show/70452188