Better layout of billing form fields with errors.
The layout got thrown off when Rails generated a .field_with_errors,
because you can't have a DIV inside of a P. These weren't paragraphs
anyway, so we took the opportunity to change each P to a .field.

We added the `table .field` style to get better padding in the credit
card/bank account table, and made `.field_with_errors` inline because
it was messing up the credit card logos. It didn't make much sense for
the presence of an error to change an input from inline to block,
anyway.

https://www.pivotaltracker.com/story/show/71423942