Don't base next billing date on actual payment date.
If a customer's card fails on their billing date, Authorize.Net may
retry the card on subsequent days. This doesn't change their annual
billing date, but when we based the next billing date on the actual
payment date, it did.

The drawback of this is, now if we're wrong about the next billing
date, receiving payment won't automatically correct it. We should
endeavor not to be wrong.