Fixed PaymentNotification#subscription bug.
It was reporting a subscription even when the notification had no
`x_subscription_id` param... because then `subscription_id` was
nil, and so `Subscription` returned the first subscription with
a null `arb_id`.

Now we don't even bother looking up the `Subscription` unless we
have a `subscription_id` to work with.

This didn't cause any actual problems, because the only thing we use
`subscription` for is to create a `Payment`, and we wouldn't create
a `Payment` for a notification with no `subscription` anyway, because
the code couldn't authenticate it. (ARB and AIM silent posts use
different authentication algorithms.) But it was confusing in the
console.