Show "please check" message only when updating billing.
For a new subscription with a blank form, the message is somewhat
condescending; of course they're going to try to get it correct.
The purpose of this message is to ask them to pay attention when
we're pre-filling the form with info that may be out of date.

We check `Organization.subscription` rather than
`@subscription.persisted` because under the hood, we may be editing a
new subscription object with info copied from the old one.

https://www.pivotaltracker.com/story/show/73218336