Upgraded to Ruby 2.1.
1.9 is no longer supported, so we need this for bug fixes and security
fixes. It isn't the latest, but it was comparatively easy to upgrade,
and it buys us some time.

Had to update DelayedJob to work with Psych instead of Syck, and
newrelic_rpm to fix every page crashing:

https://github.com/rails/rails/issues/8885

RedCloth and csv_builder were attempts to remove deprecation
warnings or something; don't remember whether they worked, but
they're minor upgrades and should be safe.