Made empty table cells sortable.
Previously an empty cell would always sink to the bottom, meaning
there was no way to see, say, "organizations that have never
logged in first."

I also toyed with having the table automatically resort when values
are changed (e.g., in the todos table), but decided this might be
confusing. The code is here, commented out, in case we decide we
want it.